requirement.txt
numpy
scikit-learn
python-Levenshtein
ollama
pypdf
pytest